pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "honey-scanner"
version = "1.0.0"
authors = [
  { name="Aldi Nova", email="your-email@example.com" },
]
description = "Web Security Assessment Tool"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Security",
]
dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "urllib3>=1.26.0",
    "lxml>=4.9.0",
    "selenium>=4.8.0",
    "cryptography>=3.4.8",
    "dnspython>=2.2.0",
    "python-whois>=0.8.0",
    "pyOpenSSL>=23.0.0",
    "fake-useragent>=1.4.0",
    "pyyaml>=6.0.1",
]

[project.urls]
"Homepage" = "https://github.com/hafourenai/webscanjust4me"
"Bug Tracker" = "https://github.com/hafourenai/webscanjust4me/issues"

[project.scripts]
honey-scanner = "honey_scanner.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["honey_scanner*"]

[tool.setuptools.package-data]
"honey_scanner.resources" = ["**/*"]
